requirements.txt
google-api-python-client
youtube-transcript-api
python-dotenv
pandas
PyYAML
